% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_object.R
\name{AddMetadata}
\alias{AddMetadata}
\title{Add metadata to the MutationObject}
\usage{
AddMetadata(obj, metadata)
}
\arguments{
\item{obj}{MutationObject.}

\item{metadata}{Sample info dataframe to add. Must have one column named
Sample_name which contain samples name identical to snp_data/snp_anno.}
}
\description{
This function is designed for add additional sample info for the object.
}
