% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_object.R
\name{AddModificationInfo}
\alias{AddModificationInfo}
\title{Add sample info to the MutationObject}
\usage{
AddModificationInfo(obj, modification_dt, type)
}
\arguments{
\item{obj}{SNPObject.}

\item{modification_dt}{Sample info dataframe to add. Must have one column named}

\item{type}{Sample info dataframe to add. Must have one column named
Sample_name which contain samples name identical to snp_data/snp_anno.}
}
\description{
This function is designed for merge multiple snp_anno and the
corresponding snp_data files, especially for multiple snp_anno and snp_data
generated from the Integrate_data() function.
}
