% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate_data.R
\name{IntegrateData}
\alias{IntegrateData}
\title{Integrate SNP result generated by snippy software}
\usage{
IntegrateData(
  file_path,
  smp_name,
  block = 50,
  nworkers = 8,
  output_dir,
  progress_bar = T,
  start_block = 1
)
}
\arguments{
\item{file_path}{a vector of file path for SNP data. The SNP data should be generated by snippy and resulted in the csv format.}

\item{smp_name}{name of each SNP sample}

\item{nworkers}{number of workers to integrate data.}

\item{progress_bar}{whether to show progress bar}

\item{chunk_size}{chunk size for each batch. This parameter is used to prevent
too many samples integrated into one batch, which will require for larger RAM.}
}
\description{
Integrate SNP result generated by snippy software
}
