% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_object.R
\name{MergeMutationObj}
\alias{MergeMutationObj}
\title{Merge multiple snp_anno and snp_data files}
\usage{
MergeMutationObj(obj_list)
}
\arguments{
\item{obj_list}{Object list.}
}
\description{
This function is designed for merge multiple snp_anno and the
SNP Object.
}
